extract spacing scrub to a separate file
